Create new methods instead code in class Program
Create new class Calculations, add 2 methods in this class. One of this
defines type of figure, and second method take this type and calculate
area for this figure. Change type of figure parameters from int to
loat. ///Insert comments for class and methods.

+++ AreaCalculations/AreaCalculations/Figure.cs
@@ -3,6 +3,6 @@ namespace AreaCalculations
 {
     public abstract class Figure
     {
-        public abstract int GetArea();
+        public abstract float GetArea();
     }
 }
